Export annotation logic
